refactor: line_break element
