+++ kraken/lib/rendering.dart
@@ -23,5 +23,5 @@ export 'src/rendering/opacity.dart';
 export 'src/rendering/transform.dart';
 export 'src/rendering/viewport.dart';
 export 'src/rendering/paragraph.dart';
-export 'src/rendering/br.dart';
+export 'src/rendering/line_break.dart';
 export 'src/rendering/image.dart';
